+++ src/main/java/me/bechberger/condensed/stats/SubStatistic.java
@@ -2,13 +2,13 @@ package me.bechberger.condensed.stats;
 
 public class SubStatistic {
     private final WriteMode mode;
-    int count = 0;
-    int bytes = 0;
-    int strings = 0;
-    int stringBytes = 0;
-    int stringLe10Bytes = 0;
-    int stringLe100Bytes = 0;
-    int stringLe1000Bytes = 0;
+    long count = 0;
+    long bytes = 0;
+    long strings = 0;
+    long stringBytes = 0;
+    long stringLe10Bytes = 0;
+    long stringLe100Bytes = 0;
+    long stringLe1000Bytes = 0;
 
     public SubStatistic(WriteMode mode) {
         this.mode = mode;
